+++ worker/src/email/index.ts
@@ -1,7 +1,7 @@
 import { Context } from "hono";
 
 import { sendMailToTelegram } from "../telegram_api";
-import { Bindings } from "../types";
+import { Bindings, HonoCustomType } from "../types";
 import { auto_reply } from "./auto_reply";
 import { trigerWebhook } from "../mails_api/webhook_settings";
 
@@ -28,7 +28,7 @@ async function email(message: ForwardableEmailMessage, env: Bindings, ctx: Execu
     // send email to telegram
     try {
         await sendMailToTelegram(
-            { env: env } as Context<{ Bindings: Bindings }>,
+            { env: env } as Context<HonoCustomType>,
             message.to, rawEmail);
     } catch (error) {
         console.log("send mail to telegram error", error);
@@ -37,7 +37,7 @@ async function email(message: ForwardableEmailMessage, env: Bindings, ctx: Execu
     // send webhook
     try {
         await trigerWebhook(
-            { env: env } as Context<{ Bindings: Bindings }>,
+            { env: env } as Context<HonoCustomType>,
             message.to, rawEmail
         );
     } catch (error) {